+++ tests/maximumflow/ridehailing_test.go
@@ -100,7 +100,7 @@ func SolveRideHailing(t *testing.T, filepath string) {
 		for v := 1; v <= n; v++ {
 			djdist[v] = math.MaxInt64
 		}
-		pq := da.NewQueryHeap[da.Index](n+1, n+1, da.ARRAY_STORAGE, true)
+		pq := da.NewQueryHeap[da.Index](uint32(n+1), uint32(n+1), da.ARRAY_STORAGE, true)
 
 		djdist[s] = 0
 		noPar := da.NewVertexEdgePair(da.INVALID_VERTEX_ID, da.INVALID_EDGE_ID, false)
